Move challenge code to rhombus_shared
