+++ rhombus-cli/Cargo.toml
@@ -25,6 +25,7 @@ rand = "0.8.5"
 ring = "0.17.8"
 directories = "5.0.1"
 markdown = "1.0.0-alpha.21"
+rhombus-shared = { path = "../rhombus-shared" }
 
 [build-dependencies]
 tonic-build = "0.12.3"
