--- rhombus-cli/build.rs
@@ -1,5 +1,0 @@
-fn main() {
-    tonic_build::configure()
-        .compile_protos(&["../proto/rhombus.proto"], &["../proto"])
-        .unwrap();
-}
