+++ rhombus-shared/src/challenges.rs
@@ -1,5 +1,5 @@
 use crate::errors::{Result, RhombusSharedError};
-use crate::grpc::proto::{Attachment, GetChallengesAdminRequest};
+use crate::proto::Attachment;
 use figment::{
     providers::{Format, Yaml},
     Figment,
@@ -96,15 +96,15 @@ pub enum AttachmentIntermediate {
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct ChallengeIntermediate {
-    stable_id: String,
-    author: String,
-    category: String,
-    description: String,
-    files: Vec<AttachmentIntermediate>,
-    flag: String,
-    name: String,
-    healthscript: Option<String>,
-    ticket_template: Option<String>,
+    pub stable_id: String,
+    pub author: String,
+    pub category: String,
+    pub description: String,
+    pub files: Vec<AttachmentIntermediate>,
+    pub flag: String,
+    pub name: String,
+    pub healthscript: Option<String>,
+    pub ticket_template: Option<String>,
 }
 
 #[derive(Clone, Debug)]
@@ -134,11 +134,9 @@ fn absolutize(base: &Path, p: &Path) -> PathBuf {
     }
 }
 
-pub async fn apply_challenges(loader_path: &Path) -> Result<()> {
-    let config: LoaderYaml = Figment::new()
-        .merge(Yaml::file_exact("loader.yaml"))
-        .extract()?;
-
+pub async fn load_challenges(
+    loader_path: &Path,
+) -> Result<BTreeMap<String, ChallengeIntermediate>> {
     let search_path = absolutize(&PathBuf::from("."), loader_path.parent().unwrap());
     let new_challenges = ChallengeYamlWalker::new(&search_path)
         .into_iter()
@@ -205,43 +203,7 @@ pub async fn apply_challenges(loader_path: &Path) -> Result<()> {
         })
         .collect::<Result<BTreeMap<_, _>>>()?;
 
-    // let response = client
-    //     .get_challenges_admin(GetChallengesAdminRequest {})
-    //     .await?
-    //     .into_inner();
-
-    // let old_challenges = response
-    //     .challenges
-    //     .into_iter()
-    //     .map(|challenge| {
-    //         (
-    //             challenge.id.clone(),
-    //             ChallengeIntermediate {
-    //                 stable_id: challenge.id,
-    //                 author: challenge.author,
-    //                 category: challenge.category,
-    //                 description: challenge.description,
-    //                 files: challenge
-    //                     .attachments
-    //                     .into_iter()
-    //                     .map(|file| AttachmentIntermediate::Literal(file))
-    //                     .collect(),
-    //                 flag: challenge.flag,
-    //                 healthscript: challenge.healthscript,
-    //                 name: challenge.name,
-    //                 ticket_template: challenge.ticket_template,
-    //             },
-    //         )
-    //     })
-    //     .collect::<BTreeMap<_, _>>();
-
-    // let difference = diff_challenges(&old_challenges, &new_challenges);
-
-    println!("=== New {:#?}\n\n\n", new_challenges);
-    // println!("=== Old {:#?}\n\n\n", old_challenges);
-    // println!("=== Dif {:#?}\n\n\n", difference);
-
-    Ok(())
+    Ok(new_challenges)
 }
 
 pub fn diff_challenges(
