+++ rhombus-shared/src/lib.rs
@@ -1,8 +1,6 @@
 pub mod challenges;
 pub mod errors;
 
-pub mod grpc {
-    pub mod proto {
-        tonic::include_proto!("rhombus");
-    }
+pub mod proto {
+    tonic::include_proto!("rhombus");
 }
